Module wigner_symbols::internal

Contents of this module are subject to change.

Structs

Step

Functions

binomial
Calculate the binomial coefficient C(n, k).

factorial
Calculate the factorial n!.

falling_factorial
Calculate the falling factorial, i.e. the product of the integers [n, n - k).

get_3tjms
Get all possible arguments of the Wigner 3-j symbol that satisfy the selection rules up to a maximum of j_max.

get_6tjs
Get all possible arguments of the Wigner 6-j symbol that satisfy the selection rules up to a maximum of j_max.

get_9tjs
Get all possible arguments of the Wigner 9-j symbol that satisfy the selection rules up to a maximum of j_max.

get_bitriangular_tjs

get_tms
Get all projection quantum numbers that lie within the multiplet of j.

get_triangular_tjs

intersect_ranges

ordering_to_i32
Reinterpret ordering as a sign:

phase

sort2

sort3

sort4

tetrahedral_sum
Calculate the symbol in the paper by L. Wei that is enclosed in square brackets:

triangle_condition
Check |j1 − j2| ≤ j3 ≤ j1 + j2 and j1 + j2 + j3 ∈ ℤ.

triangular_factor
Calculate the triangular factor:

triangular_factor_raw
Calculate ja! jb! jc! / jd!.

wigner_3jm_raw
Calculate the Wigner 3-jm symbol times (−1) ^ (j1 − j2 − m3). The selection rules are not checked.

wigner_3jm_raw_c
Calculate the Wigner 3-jm symbol times (−1) ^ (j1 − j2 − m3).

wigner_6j_raw
Calculate the Wigner 6-j symbol. The selection rules are not checked.

wigner_9j_raw
Calculate the Wigner 9-j symbol. The selection rules are not checked.