Struct wigner_symbols::SignedSqrt

pub struct SignedSqrt(pub Rational);

Signed square root of a rational number

This represents a mathematical expression of the form s √(n / d) where

s is a sign (-1, 0, or +1),

n is a nonnegative numerator, and

d is a positive denominator.

Internally, it is represented by the rational number s × n / d.

This can be converted to a floating-point number via f64::from(…).

Defaults to zero.

Methods

impl SignedSqrt

pub fn new(c: Integer, r: Rational) -> Self

Construct a SignedSqrt equal to c √r.

pub fn sign(&self) -> Ordering

Equivalent to self.cmp(&Self::from(0)).

pub fn sq(self) -> Rational

Returns the square of the expression.

pub fn signed_sq(self) -> Rational

Returns the square of the expression, but with the sign adjusted to match the sign of the original expression.